backend/requirements.txt
ccxt
python-dotenv
feedparser
flask
flask_cors
matplotlib
numpy
pandas
praw
prometheus_client
pytrends
requests
vaderSentiment
websocket-client
werkzeug
PyYAML
gunicorn
pydantic

